+++ NEWS.md
@@ -1,5 +1,5 @@
 # crunch 1.27.5 (Development version)
-* You can set the `weight` and `filter` to be the same on each slide of a deck.
+* You can set the weight and filter to the same value for every slide in a deck by assigning a `weight` or `filter` on the deck object.
 
 # crunch 1.27.4
 * Fixes to pkgdown documentation website